Adding base4kidsIsDisabled attribute
The base4kidsIsDisabled attribute indicates whether a base4kids object is
disabled (but still active) or not.

In contrast to the base4kidsIsBlocked attribute, the base4kidsIsDisabled
attribute must be used, if the user has been marked as delete on the master
database. Whereas the base4kidsIsBlocked attribute must be set if the user has
been blocked manually through an administrative user.